Move sub-locations fetch/create to SubLocationService
- I had originally created them under SeedBankService (since seedbanks alone referred to sublocations)
- move these functions into the newly created SubLocationService and update all callers
- plan is to use these functions in the nurseries details view updates
