+++ src/services/SubLocationService.ts
@@ -5,10 +5,14 @@ import HttpService, { Response } from './HttpService';
 import { getPromisesResponse } from './utils';
 
 /**
- * Seed bank related services
+ * Sub-locations CRUD
  */
 
+const SUB_LOCATIONS_ENDPOINT = '/api/v1/facilities/{facilityId}/subLocations';
 const SUB_LOCATION_ENDPOINT = '/api/v1/facilities/{facilityId}/subLocations/{subLocationId}';
+
+type StorageLocationsResponsePayload =
+  paths[typeof SUB_LOCATIONS_ENDPOINT]['get']['responses'][200]['content']['application/json'];
 
 type SubLocationResponsePayload =
   paths[typeof SUB_LOCATION_ENDPOINT]['get']['responses'][200]['content']['application/json'];
@@ -16,6 +20,11 @@ type SubLocationResponsePayload =
 /**
  * exported types
  */
+
+export type SubLocationsData = {
+  subLocations: SubLocation[];
+};
+export type SubLocationsResponse = Response & SubLocationsData;
 
 export type SubLocationData = {
   subLocation?: SubLocation;
@@ -25,7 +34,33 @@ export type SubLocationResponse = Response & SubLocationData;
 export type SubLocationUpdateRequestBody =
   paths[typeof SUB_LOCATION_ENDPOINT]['put']['requestBody']['content']['application/json'];
 
+const httpSubLocations = HttpService.root(SUB_LOCATIONS_ENDPOINT);
 const httpStorageLocation = HttpService.root(SUB_LOCATION_ENDPOINT);
+
+/**
+ * Create a single sub-location
+ */
+const createSubLocation = async (facilityId: number, name: string): Promise<SubLocationResponse> => {
+  const response: Response = await httpSubLocations.post({
+    urlReplacements: {
+      '{facilityId}': facilityId.toString(),
+    },
+    entity: { name },
+  });
+
+  return {
+    ...response,
+    subLocation: response.data?.subLocation,
+  };
+};
+
+/**
+ * Create one or more sub-locations
+ */
+const createSubLocations = async (facilityId: number, names: string[]): Promise<(SubLocationResponse | null)[]> => {
+  const promises = names.map((name) => createSubLocation(facilityId, name));
+  return getPromisesResponse<SubLocationResponse>(promises);
+};
 
 /**
  * Get a single sub-location
@@ -39,6 +74,22 @@ const getSubLocation = async (facilityId: number, locationId: number): Promise<S
       },
     },
     (data) => ({ subLocation: data?.subLocation })
+  );
+
+  return response;
+};
+
+/*
+ * Returns all the sub-locations associated with a given facility or null if the API call failed.
+ */
+const getSubLocations = async (facilityId: number): Promise<SubLocationsResponse> => {
+  const response: SubLocationsResponse = await httpSubLocations.get<StorageLocationsResponsePayload, SubLocationsData>(
+    {
+      urlReplacements: {
+        '{facilityId}': facilityId.toString(),
+      },
+    },
+    (data) => ({ subLocations: data?.subLocations ?? [] })
   );
 
   return response;
@@ -92,7 +143,10 @@ const deleteSubLocations = async (facilityId: number, locationIds: number[]): Pr
  * Exported functions
  */
 const SubLocationService = {
+  createSubLocation,
+  createSubLocations,
   getSubLocation,
+  getSubLocations,
   updateSubLocation,
   deleteSubLocation,
   updateSubLocations,
